Started rewriting game objects.
