+++ python_src/hades/game_objects/characteristics.py
@@ -1,0 +1,2 @@
+"""Manages the different characteristics available."""
+from __future__ import annotations
