+++ python_src/hades/game_objects/objects.py
@@ -1,0 +1,46 @@
+"""Manages the base classes used by all game objects."""
+from __future__ import annotations
+
+# Builtin
+import logging
+
+# Pip
+import arcade
+
+# Custom
+from hades.constants.game_objects import SPRITE_SCALE
+from hades.textures import grid_pos_to_pixel
+
+__all__ = ("GameObject",)
+
+# Get the logger
+logger = logging.getLogger(__name__)
+
+
+class GameObject(arcade.Sprite):
+    """The base class for all game objects.
+
+    Attributes
+    ----------
+        center_x: float
+            The x position of the object on the screen.
+        center_y: float
+            The y position of the object on the screen.
+    """
+
+    def __init__(self, x: int, y: int) -> None:
+        """Initialise the object.
+
+        Parameters
+        ----------
+        x: int
+            The x position of the object in the game map.
+        y: int
+            The y position of the object in the game map.
+        """
+        super().__init__(scale=SPRITE_SCALE)
+        self.center_x, self.center_y = grid_pos_to_pixel(x, y)
+
+    def __repr__(self) -> str:
+        """Return a human-readable representation of this object."""
+        return f"<GameObject (Position={self.position})>"
