+++ python_src/hades/game_objects_OLD/attributes.py
