+++ python_src/hades/melee_shader.py
@@ -13,7 +13,7 @@ from hades.constants.game_objects import MELEE_RESOLUTION, SPRITE_SIZE
 if TYPE_CHECKING:
     from arcade import ArcadeContext
     from arcade.gl import Buffer, Framebuffer, Program, Query
-    from hades.game_objects.enemies import Enemy
+    from hades.game_objects_OLD.enemies import Enemy
     from hades.views.game import Game
 
 __all__ = ("MeleeShader",)
