+++ tests/game_objects/__init__.py
@@ -1,0 +1,1 @@
+"""Runs automated tests on all files in game_objects/."""
