+++ tests/game_objects/test_characteristics.py
@@ -1,0 +1,4 @@
+"""Tests all functions in game_objects/characteristics.py."""
+from __future__ import annotations
+
+__all__ = ()
