+++ tests/game_objects/test_components.py
@@ -1,0 +1,4 @@
+"""Tests all functions in game_objects/components.py."""
+from __future__ import annotations
+
+__all__ = ()
