+++ tests/game_objects/test_entities.py
@@ -1,0 +1,4 @@
+"""Tests all functions in game_objects/entities.py."""
+from __future__ import annotations
+
+__all__ = ()
