+++ tests/game_objects/test_objects.py
@@ -1,0 +1,4 @@
+"""Tests all functions in game_objects/objects.py."""
+from __future__ import annotations
+
+__all__ = ()
