+++ tests/game_objects/test_tiles.py
@@ -1,0 +1,4 @@
+"""Tests all functions in game_objects/tiles.py."""
+from __future__ import annotations
+
+__all__ = ()
